requirements.txt
MetaTrader5
pandas
numpy
pyyaml
matplotlib
